requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distro-info==1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.20.1
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0